% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_CRAN.R
\name{build_tag_vector}
\alias{build_tag_vector}
\title{Build Final Tag Vector}
\usage{
build_tag_vector(tags, additional_tags, use_defaults)
}
\description{
Internal helper to construct tag vector from various inputs
}
\keyword{internal}
