% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_tool}
\alias{classify_tool}
\title{Classify Tool by Name}
\usage{
classify_tool(name)
}
\arguments{
\item{name}{Character string of the tool name}
}
\value{
Character string of the tool category
}
\description{
Categorizes a bioinformatics tool based on its name
}
\examples{
classify_tool("FastQC")
classify_tool("QIIME2")
}
