% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_mock_catalog}
\alias{create_mock_catalog}
\title{Create a mock catalog data frame}
\usage{
create_mock_catalog(n_cran = 2, n_bioc = 2, n_github = 2)
}
\arguments{
\item{n_cran}{Number of CRAN packages}

\item{n_bioc}{Number of Bioconductor packages}

\item{n_github}{Number of GitHub repos}
}
\value{
Data frame of mock catalog
}
\description{
Create a mock catalog data frame
}
