% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_tags.R
\name{default_bacterial_tags}
\alias{default_bacterial_tags}
\title{Default Tags for Bacterial Research}
\format{
A character vector with 10 elements
}
\usage{
default_bacterial_tags
}
\description{
A character vector of default search tags for bacterial and microbiome research
}
