% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_Github.R
\name{extract_licenses}
\alias{extract_licenses}
\title{Extract License Information from GitHub API Response}
\usage{
extract_licenses(items)
}
\description{
Extract License Information from GitHub API Response
}
\keyword{internal}
