% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_Github.R
\name{get_doi_from_citation_file}
\alias{get_doi_from_citation_file}
\title{Get DOI from CITATION.cff File}
\usage{
get_doi_from_citation_file(owner, repo, headers, cache_env)
}
\description{
Get DOI from CITATION.cff File
}
\keyword{internal}
