% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tag_pattern.R, R/utils.R
\name{make_tag_pattern}
\alias{make_tag_pattern}
\title{make_tag_pattern
The pattern produced by the function makes it easy to be used in functions like
grep(), str_detect(), or other string matching operations in R to find any
occurrences of these keywords.}
\usage{
make_tag_pattern(tags)

make_tag_pattern(tags)
}
\arguments{
\item{tags}{Character vector of tags}
}
\value{
A regex string that matches any of the tags.

A regex pattern string with tags separated by |
}
\description{
make_tag_pattern
The pattern produced by the function makes it easy to be used in functions like
grep(), str_detect(), or other string matching operations in R to find any
occurrences of these keywords.

Create Tag Pattern for Regex Matching
}
\examples{
Input: ("bacteria", "microbiome", "antibiotic") or c("bacteria", "microbiome", "antibiotic")
pattern <- make_tag_pattern("bacteria", "microbiome", "microbiomes")
}
\keyword{internal}
