% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_Bioconductor.R
\name{query_bioc_by_tags}
\alias{query_bioc_by_tags}
\title{Query Bioconductor Packages by Tags}
\usage{
query_bioc_by_tags(tags = NULL, additional_tags = NULL, use_defaults = TRUE)
}
\arguments{
\item{tags}{Character vector of search tags (default: default_bacterial_tags)}

\item{additional_tags}{Character vector of additional tags to append to the default tags (default: NULL)}

\item{use_defaults}{Logical, whether to include default_bacterial_tags (default: TRUE)}
}
\value{
List of package information
}
\description{
Search Bioconductor repository for packages matching specified tags
}
\examples{
\dontrun{
bioc_tools <- query_bioc_by_tags()
bioc_tools <- query_bioc_by_tags(additional_tags = c("sequencing", "assembly"))
}
}
