% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tool_list_to_df}
\alias{tool_list_to_df}
\title{Convert Tool List to Data Frame}
\usage{
tool_list_to_df(tool_list)
}
\arguments{
\item{tool_list}{List of tools with nested structure}
}
\value{
A data frame with tool information
}
\description{
Convert Tool List to Data Frame
}
\keyword{internal}
